Skip looking for holes when count(id) == max(id), bugfix
This change avoids looking for holes in group numerical ids when
the number of ids matches the maximum id. We can do that since we
use sequential integer ids starting from 1. Also, a bug was found
and corrected related to the expected behaviour when a new group
invitation is used by a user that is already part of a different
group.