Also clean existing groups when generating new invitations
We previously added to the invitation generation script that it
would remove all previously existing invitations but we forgot to
remove the existing groups. It's safe to assume that if we are
generating invitations it means that a new yearly event is going
to occur. We could always remove that line if for some reason
we only needed to add some extra invitations for the current
event.