Fix sessions not working and other minor changes
Apparently in one of the small changes I did a while ago I broke
handling uids through sessions and I've spent a while trying to
understand how this was now broken if it worked fine for a long
time. It turns out that I added a new variable that was both
useless and removing the ability of getting uids from previously
used invitations.