Handle event/now times (and their timezones) properly
This change fixes a long due bug that made time work half the year
since it was assumed UTC was 2 hours behind Europe/Madrid's time,
but that is not true for 6 months every year. Since testing is
being made during the half that it is not right it was specially
annoying. Now we handle datetimes and zones just right, as it
should be.